Fix cognitive complexity plugin else if increments and nesting
Else if statements were causing a nesting increment (increasing the cognitive complexity by the current nesting level) when they should not have. Only the first if in the chain has the nesting penalty.

A chain of else if's was erroneously decreasing the nesting level by the number of else if statements.